+++ src/app/behavioral/components/InterviewPanel.tsx
@@ -2,10 +2,196 @@
 // Main panel component that orchestrates the interview experience.
 // Manages state between question display, user responses, and feedback.
 
+import { useState, useRef, useEffect } from 'react';
+import QuestionDisplay from './QuestionDisplay';
+
+interface Message {
+  id: string;
+  role: 'user' | 'assistant';
+  content: string;
+  timestamp: Date;
+}
+
 export default function InterviewPanel() {
+  const initialQuestion = "Hello! I'm your AI interviewer for today's behavioral interview. I'll be asking you questions about your past experiences and how you handle various situations. Are you ready to begin?";
+  const [messages, setMessages] = useState<Message[]>([
+    {
+      id: '1',
+      role: 'assistant',
+      content: initialQuestion,
+      timestamp: new Date(),
+    },
+  ]);
+  const [inputValue, setInputValue] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+  const messagesEndRef = useRef<HTMLDivElement>(null);
+  const textareaRef = useRef<HTMLTextAreaElement>(null);
+
+  useEffect(() => {
+    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+  }, [messages]);
+
+  useEffect(() => {
+    if (textareaRef.current) {
+      textareaRef.current.style.height = 'auto';
+      textareaRef.current.style.height = textareaRef.current.scrollHeight + 'px';
+    }
+  }, [inputValue]);
+
+  const handleSendMessage = async () => {
+    if (!inputValue.trim() || isLoading) return;
+
+    const userMessage: Message = {
+      id: Date.now().toString(),
+      role: 'user',
+      content: inputValue.trim(),
+      timestamp: new Date(),
+    };
+
+    setMessages((prev) => [...prev, userMessage]);
+    setInputValue('');
+    setIsLoading(true);
+
+    try {
+      const response = await fetch('/api/gemini', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          messages: [...messages, userMessage].map((msg) => ({
+            role: msg.role,
+            content: msg.content,
+          })),
+          interviewType: 'behavioral',
+        }),
+      });
+
+      if (!response.ok) {
+        throw new Error('Failed to get response');
+      }
+
+      const data = await response.json();
+      
+      const assistantMessage: Message = {
+        id: (Date.now() + 1).toString(),
+        role: 'assistant',
+        content: data.response || 'I apologize, but I encountered an error. Could you please try again?',
+        timestamp: new Date(),
+      };
+
+      setMessages((prev) => [...prev, assistantMessage]);
+    } catch (error) {
+      console.error('Error sending message:', error);
+      const errorMessage: Message = {
+        id: (Date.now() + 1).toString(),
+        role: 'assistant',
+        content: 'I apologize, but I\'m having trouble connecting right now. Please try again.',
+        timestamp: new Date(),
+      };
+      setMessages((prev) => [...prev, errorMessage]);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  const handleKeyPress = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if (e.key === 'Enter' && !e.shiftKey) {
+      e.preventDefault();
+      handleSendMessage();
+    }
+  };
+
   return (
-    <div>
-      {/* Interview panel layout will be implemented here */}
+    <div className="min-h-screen bg-gray-50">
+      {/* Header */}
+      <div className="bg-white border-b border-gray-200 shadow-sm">
+        <div className="max-w-3xl mx-auto px-6 py-5">
+          <h1 className="text-xl font-semibold text-black">
+            Behavioral Interview
+          </h1>
+          <p className="text-xs text-gray-500 mt-1">
+            Practice with an AI interviewer
+          </p>
+        </div>
+      </div>
+
+      {/* Chat Container */}
+      <div className="max-w-3xl mx-auto px-6 py-8 h-[calc(100vh-160px)] flex flex-col">
+        {/* Question Display */}
+  <QuestionDisplay currentMessage="" />
+        {/* Messages */}
+        <div className="flex-1 overflow-y-auto mb-6 space-y-5">
+          {messages.map((message) => (
+            <div
+              key={message.id}
+              className={`flex ${
+                message.role === 'user' ? 'justify-end' : 'justify-start'
+              }`}
+            >
+              <div
+                className={`max-w-[75%] rounded-lg ${
+                  message.role === 'user'
+                    ? 'bg-black text-white px-4 py-3'
+                    : 'bg-white border border-gray-200 text-black px-4 py-3'
+                }`}
+              >
+                <p className="text-sm leading-relaxed whitespace-pre-wrap">{message.content}</p>
+                <span
+                  className={`text-xs mt-2 block ${
+                    message.role === 'user'
+                      ? 'text-gray-300'
+                      : 'text-gray-400'
+                  }`}
+                >
+                  {message.timestamp.toLocaleTimeString([], {
+                    hour: '2-digit',
+                    minute: '2-digit',
+                  })}
+                </span>
+              </div>
+            </div>
+          ))}
+          {isLoading && (
+            <div className="flex justify-start">
+              <div className="bg-white border border-gray-200 rounded-lg px-4 py-3">
+                <div className="flex space-x-1">
+                  <div className="w-1.5 h-1.5 bg-gray-600 rounded-full animate-bounce"></div>
+                  <div className="w-1.5 h-1.5 bg-gray-600 rounded-full animate-bounce delay-100"></div>
+                  <div className="w-1.5 h-1.5 bg-gray-600 rounded-full animate-bounce delay-200"></div>
+                </div>
+              </div>
+            </div>
+          )}
+          <div ref={messagesEndRef} />
+        </div>
+
+        {/* Input Area */}
+        <div className="bg-white border border-gray-200 rounded-lg shadow-sm p-4">
+          <div className="flex items-center gap-3">
+            <div className="flex flex-1 items-center h-10">
+              <textarea
+                ref={textareaRef}
+                value={inputValue}
+                onChange={(e) => setInputValue(e.target.value)}
+                onKeyPress={handleKeyPress}
+                placeholder="Type your response..."
+                className="w-full resize-none h-full px-3 py-2 bg-transparent border-none focus:outline-none text-black placeholder-gray-400 text-sm flex items-center"
+                rows={1}
+                disabled={isLoading}
+                style={{ minHeight: '40px' }}
+              />
+            </div>
+            <button
+              onClick={handleSendMessage}
+              disabled={!inputValue.trim() || isLoading}
+              className="h-10 px-5 py-2 bg-black text-white text-sm font-medium rounded hover:bg-gray-800 focus:outline-none disabled:bg-gray-300 disabled:cursor-not-allowed transition-colors align-middle flex items-center"
+            >
+              Send
+            </button>
+          </div>
+        </div>
+      </div>
     </div>
   );
 }
