+++ src/app/behavioral/components/QuestionDisplay.tsx
@@ -2,10 +2,16 @@
 // Component responsible for displaying interview questions from the AI.
 // Handles text-to-speech integration and question progression.
 
-export default function QuestionDisplay() {
+interface QuestionDisplayProps {
+  currentMessage: string;
+}
+
+export default function QuestionDisplay({ currentMessage }: QuestionDisplayProps) {
   return (
-    <div>
-      {/* Question display and TTS controls will be implemented here */}
+    <div className="mb-6">
+      <div className="text-lg font-medium text-black">
+        {currentMessage}
+      </div>
     </div>
   );
 }
